+++ backend/database/database_util.go
@@ -1,17 +1,18 @@
 // utility functions for questions api
-package main
+package database
 
 import (
 	"context"
 	"fmt"
 
 	"go.mongodb.org/mongo-driver/bson"
+	"peerprep/common"
 )
 
 // used to check if a question already exists in the database.
-func (db *QuestionDB) QuestionExists(question *Question) bool {
+func (db *QuestionDB) QuestionExists(question *common.Question) bool {
 	filter := bson.D{bson.E{Key: "title", Value: question.Title}}
-	err := db.questions.FindOne(context.Background(), filter).Decode(&Question{}) //FindOne() returns error if no document is found
+	err := db.questions.FindOne(context.Background(), filter).Decode(&common.Question{}) //FindOne() returns error if no document is found
 	return err == nil
 }
 
@@ -31,7 +32,7 @@ func (db *QuestionDB) FindNextQuestionId() int {
 
 // used to increment the next question ID to be used.
 // since the collection is capped at one document, inserting a new document will replace the old one.
-func (db *QuestionDB) IncrementNextQuestionId(nextId int, logger *Logger) {
+func (db *QuestionDB) IncrementNextQuestionId(nextId int, logger *common.Logger) {
 	var err error
 	if _, err = db.nextId.InsertOne(context.Background(), bson.D{bson.E{Key: "next", Value: nextId}}); err != nil {
 		logger.Log.Error("Error incrementing next question ID: ", err)
@@ -43,8 +44,8 @@ func (db *QuestionDB) IncrementNextQuestionId(nextId int, logger *Logger) {
 
 // used to check if a question being replaced will cause duplicates in the database
 
-func (db *QuestionDB) QuestionExistsExceptId(question *Question) bool {
+func (db *QuestionDB) QuestionExistsExceptId(question *common.Question) bool {
 	filter := bson.D{bson.E{Key: "title", Value: question.Title}, bson.E{Key: "id", Value: bson.D{bson.E{Key: "$ne", Value: question.ID}}}}
-	err := db.questions.FindOne(context.Background(), filter).Decode(&Question{}) //FindOne() returns error if no document is found
+	err := db.questions.FindOne(context.Background(), filter).Decode(&common.Question{}) //FindOne() returns error if no document is found
 	return err == nil
 }
